requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
Flask==2.1.1
Flask-Login==0.6.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
motor==2.5.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.6
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
